Un-spaghetti-ize some logic via fewer #ifdefs
Always call prepare_pages when about to use one or more pages that
were previously unused. This unifies the flow regarding lazy commit
or zero-fill. I still need to incorporate darwin-jit re-mapping.

+++ src/runtime/monitor.c
@@ -129,7 +129,7 @@ static void checked_write(char *section, struct filewriter* writer, void* buf, l
     if (write(writer->fd, label, sizeof label) != sizeof label
         || (wrote = write(writer->fd, buf, nbytes)) != nbytes)
         lose("short write, errno=%d", errno);
-    fprintf(stderr, "%s: %lx bytes\n", section, nbytes + sizeof label);
+    fprintf(stderr, "%s: %lx bytes\n", section, (long)(nbytes + sizeof label));
     maybe_show_contents(section, buf, nbytes);
     writer->total += nbytes + sizeof label;
 }
